Fix warning in Rails 4
And keep same behavior in Rails 5. I don't think the fact that this
columns allow null or not could cause any problems, but I guess it's
better to keep the same behavior we've always had.

The warning is

```
DEPRECATION WARNING: `#timestamps` was called without specifying an
option for `null`. In Rails 5, this behavior will change to
`null: false`. You should manually specify `null: true` to prevent the
behavior of your existing migrations from changing.
```